split files in the macro project
