+++ macros/src/main/scala/com/unstablebuild/slime/AnnotatedInstance.scala
@@ -1,0 +1,7 @@
+package com.unstablebuild.slime
+
+case class AnnotatedInstance[T](instance: T, encoder: TypeEncoder[T]) {
+
+  def encoded: Seq[(String, Value)] = encoder.encode(instance)
+
+}
