+++ macros/src/main/scala/com/unstablebuild/slime/TypeEncoder.scala
@@ -1,0 +1,7 @@
+package com.unstablebuild.slime
+
+trait TypeEncoder[-T] {
+
+  def encode(instance: T): Seq[(String, Value)]
+
+}
